+++ src/main/java/ai/mcpdirect/studio/service/AIPortServerResponse.java
@@ -1,6 +1,0 @@
-package ai.mcpdirect.studio.service;
-
-import appnet.hstp.SimpleServiceResponseMessage;
-
-public class AIPortServerResponse<T> extends SimpleServiceResponseMessage<T> {
-}